Refactor and util.condition function
